Save changes
bug not fixed: the favorite parameter in the Course class gets updated
for the course (object) inside the setFavorite function but is set to
its default value false before any other function uses the parameter.
Need to figure out why is the value changing back to false.

Also, reloading the page messes everything up. works fine for
restarting the app.